==> code/2_Fine_Tuning.ipynb <==
# 0. Intro

The Sentiment Analysis process needs the following steps:

- Feature extraction
- Modeling

#### Feature Extraction

This phase is based on statistical bag-of-words methods (CountVectorizer, TfidfVectorizer)

#### Modeling

The models involved are Traditional Machine Learning ones: DecisionTree, RandomForest, GradientBoosting + VotingClassifier


Both these points entail some parameters. In order to maximize results some of these parameters are going to be tuned. This operation represents the core of this notebook 



# 1. Packages & Basic Settings

import numpy as np
import pandas as pd
import os
import pickle

from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, HistGradientBoostingClassifier, VotingClassifier
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.model_selection import TimeSeriesSplit
from sklearn.metrics import make_scorer, matthews_corrcoef
from sklearn.pipeline import Pipeline

from pipe_optimization import NlpPipeBayesSearch

data_path = '../data'

seed = 0


# 2. Data Import

preprocessed_df_filename = 'df_preprocessed.parquet'

df = pd.read_parquet(os.path.join(data_path, 'intermediate', preprocessed_df_filename))

# 3. Data Segmentation

overall_test_size = 0.2

split_point = int(round(len(df)*(1-overall_test_size)))


text_train=df['clean_text'].values[:split_point]
y_bin_train=df['binary_label'].values[:split_point]
y_ter_train=df['binary_label'].values[:split_point]

# 4. Parameters Search

Vectorizers hyperparameters:
- max features -> number of tokens (most frequent) to keep as features 

Models hyperparameters:

DecisionTree: [max_depth, min_samples_leaf]

RandomForest: [n_estimators, max_depth, max_features]

GB: [max_iter, max_depth, learning_rate]


#### Search method: BayesSearch
#### Evaluation metric: Matthews correlation coefficient (MCC)

## 4.1. Search params initialization

#### Vectorizers fixed params and search space

vects_space = {'cv':{'object':CountVectorizer(ngram_range=(1,4)), 
                     'space':{'max_features':(2**8,2**11)}},
        'tfidf':{'object':TfidfVectorizer(ngram_range=(1,4)), 
                 'space':{'max_features':(2**8,2**11)}}}

#### Models fixed params and search space

models_space = {'DT':{'object':DecisionTreeClassifier(random_state=seed), 
                      'space':{'max_depth':(2,4), 'min_samples_leaf':(2**3, 2**5)}},
    'RF':{'object':RandomForestClassifier(random_state=seed), 
          'space':{'max_depth':(2,4), 'n_estimators':(2**5, 2**8), 'max_features':(1/2**6,1/2**2)}},
    'HGB':{'object':HistGradientBoostingClassifier(random_state=seed), 
           'space':{'max_depth':(2,4), 'max_iter':(2**5, 2**8), 'learning_rate':(1/2**9,1/2**3)}}}

#### Scorer

mcc = make_scorer(matthews_corrcoef)

#### CV object

n_cv = 3
cv_test_size = 400

ts_split = TimeSeriesSplit(n_splits=n_cv, test_size=cv_test_size)

#### N jobs

nj = os.cpu_count() - 1


## 4.2. Bayes Search

n_iterations = 16

binary_pipe_bay_search = NlpPipeBayesSearch(vects_dict=vects_space,
                                     clfs_dict=models_space,
                                     cv_object=ts_split,
                                     n_iter=n_iterations,
                                     random_state=seed,
                                     n_jobs=nj,
                                     scoring=mcc,
                                     std_penalty=True)


binary_pipe_bay_search.search(text_train, y_bin_train)

ternary_pipe_bay_search = NlpPipeBayesSearch(vects_dict=vects_space,
                                     clfs_dict=models_space,
                                     cv_object=ts_split,
                                     n_iter=n_iterations,
                                     random_state=seed,
                                     n_jobs=nj,
                                     scoring=mcc,
                                     std_penalty=True)


ternary_pipe_bay_search.search(text_train, y_ter_train)

# 5. Results Storage

## 5.1. Best pipelines

pipelines_instructions = {}

pipelines_instructions['binary'] = binary_pipe_bay_search.pipelines_instructions.copy()

pipelines_instructions['ternary'] = ternary_pipe_bay_search.pipelines_instructions.copy()

Adding Voting Classifiers, one for each vectorizer. Their estimators are the best pipelines involving other models

for vec in vects_space.keys():
    
    pipelines_instructions['binary'][(vec, 'VC')] = [('clf', VotingClassifier(estimators = [(vzer.lower()+'_'+mdl.lower(), Pipeline(pipe)) for (vzer,mdl), pipe in pipelines_instructions['binary'].items() if vzer==vec]))]

    pipelines_instructions['ternary'][(vec, 'VC')] = [('clf', VotingClassifier(estimators = [(vzer.lower()+'_'+mdl.lower(), Pipeline(pipe)) for (vzer,mdl), pipe in pipelines_instructions['ternary'].items() if vzer==vec]))]
    



## 5.2. Cross Validation Results

cv_res = {}

cv_res['binary'] = binary_pipe_bay_search.cv_results.copy()

cv_res['ternary'] = ternary_pipe_bay_search.cv_results.copy()

# 6. Export

pickle.dump(pipelines_instructions, open(os.path.join(data_path, 'output', 'NLP_FSA_pipelines_instructions.pkl'), 'wb'))

pickle.dump(cv_res, open(os.path.join(data_path, 'output', 'NLP_FSA_bcv_results.pkl'), 'wb'))
