==> code/3_Evaluate.ipynb <==
# 0. Intro

Notebook to deploy and evaluate the optimized pipelines

# 1. Packages & Basic Settings

import numpy as np
import pandas as pd
import os
import pickle
import matplotlib.pyplot as plt
import shap
from sklearn.pipeline import Pipeline
from sklearn.ensemble import RandomForestClassifier, HistGradientBoostingClassifier
from sklearn.tree import DecisionTreeClassifier

from evaluation import CustomEval

data_path = '../data'

# 2. Data

& Pipelines instructions

preprocessed_df_filename = 'df_preprocessed.parquet'

df = pd.read_parquet(os.path.join(data_path, 'intermediate', preprocessed_df_filename))

pipelines_instructions = pickle.load(open(os.path.join(data_path, 'output', 'NLP_FSA_pipelines_instructions.pkl'), 'rb'))

# 3. Data Segmentation

overall_test_size = 0.2

split_point = int(round(len(df)*(1-overall_test_size)))

df_train = df.iloc[:split_point].copy()

df_test = df.iloc[split_point:].copy()


# 4. Deploy & Evaluate 

test_rets = df_test['return'].values

selected_metrics = ['MCC','Accuracy','F1','Roc_Auc','PnL_sum','MDD']



extended_names = {'cv': 'CountVectorizer',
                  'tfidf': 'TfidfVectorizer',
                  'binary':'Binary',
                  'ternary':'Ternary',
                  'DT': 'DecisionTree',
                  'RF': 'RandomForest',
                  'HGB': 'GradientBoosting',
                  'VC': 'VotingClassifier'}

all_results = []

for clf, pipes in pipelines_instructions.items():

    for (vect,model), pipe_steps in pipes.items():

        pipe_points = {'Classification':extended_names[clf], 'Vectorizer':extended_names[vect], 'Model':extended_names[model]}

        pipe = Pipeline(pipe_steps)

        pipe.fit(df_train['clean_text'], df_train[f'{clf}_label'])

        y_pred = pipe.predict(df_test['clean_text'])

        computed_metrics = CustomEval(clf=clf, metrics = selected_metrics).eval(df_test[f'{clf}_label'].values, y_pred, test_rets)

        results_store = dict(**pipe_points, **computed_metrics) 

        all_results.append(results_store)


df_res = pd.DataFrame(all_results).sort_values(by=['Classification','Vectorizer']).round(4)

df_res
